Move `zbc` out of `c` and under `zce`
