+++ crates/execution/ab-riscv-act4-runner/src/abundance_rv32i_max/instruction.rs
@@ -32,8 +32,8 @@ use ab_riscv_primitives::instructions::rv32::b::zbb::Rv32ZbbInstruction;
 use ab_riscv_primitives::instructions::rv32::b::zbc::Rv32ZbcInstruction;
 use ab_riscv_primitives::instructions::rv32::b::zbs::Rv32ZbsInstruction;
 use ab_riscv_primitives::instructions::rv32::c::zca::Rv32ZcaInstruction;
-use ab_riscv_primitives::instructions::rv32::c::zcb::Rv32ZcbInstruction;
 use ab_riscv_primitives::instructions::rv32::m::Rv32MInstruction;
+use ab_riscv_primitives::instructions::rv32::zce::zcb::Rv32ZcbInstruction;
 use ab_riscv_primitives::instructions::rv32::zk::zbkb::Rv32ZbkbInstruction;
 use ab_riscv_primitives::instructions::rv32::zk::zbkx::Rv32ZbkxInstruction;
 use ab_riscv_primitives::instructions::rv32::zk::zkn::zknd::{Rv32AesBs, Rv32ZkndInstruction};
