+++ crates/execution/ab-riscv-act4-runner/src/abundance_rv64i_max/instruction.rs
@@ -32,8 +32,8 @@ use ab_riscv_primitives::instructions::rv64::b::zbb::Rv64ZbbInstruction;
 use ab_riscv_primitives::instructions::rv64::b::zbc::Rv64ZbcInstruction;
 use ab_riscv_primitives::instructions::rv64::b::zbs::Rv64ZbsInstruction;
 use ab_riscv_primitives::instructions::rv64::c::zca::Rv64ZcaInstruction;
-use ab_riscv_primitives::instructions::rv64::c::zcb::Rv64ZcbInstruction;
 use ab_riscv_primitives::instructions::rv64::m::Rv64MInstruction;
+use ab_riscv_primitives::instructions::rv64::zce::zcb::Rv64ZcbInstruction;
 use ab_riscv_primitives::instructions::rv64::zk::zbkb::Rv64ZbkbInstruction;
 use ab_riscv_primitives::instructions::rv64::zk::zbkx::Rv64ZbkxInstruction;
 use ab_riscv_primitives::instructions::rv64::zk::zkn::zknd::{Rv64ZkndInstruction, Rv64ZkndKsRnum};
