+++ crates/execution/ab-riscv-interpreter/src/rv32/c.rs
@@ -1,4 +1,3 @@
 //! RV32 C extension
 
 pub mod zca;
-pub mod zcb;
