+++ crates/execution/ab-riscv-interpreter/src/rv32/zce.rs
@@ -1,0 +1,3 @@
+//! RV32 Zce extension
+
+pub mod zcb;
