+++ crates/execution/ab-riscv-interpreter/src/rv32/zce/zcb.rs
@@ -1,4 +1,6 @@
-//! Zcb compressed instruction execution
+//! Zcb compressed instruction execution (RV32)
+//!
+//! C.ZEXT.W is absent in RV32 - the enum has no such variant.
 
 #[cfg(test)]
 mod tests;
@@ -8,7 +10,7 @@ use crate::{
     SystemInstructionHandler, VirtualMemory,
 };
 use ab_riscv_macros::instruction_execution;
-use ab_riscv_primitives::instructions::rv64::c::zcb::Rv64ZcbInstruction;
+use ab_riscv_primitives::instructions::rv32::zce::zcb::Rv32ZcbInstruction;
 use ab_riscv_primitives::registers::general_purpose::Register;
 use core::ops::ControlFlow;
 
@@ -17,9 +19,9 @@ impl<Reg, ExtState, Memory, PC, InstructionHandler, CustomError>
     ExecutableInstruction<
         InterpreterState<Reg, ExtState, Memory, PC, InstructionHandler, CustomError>,
         CustomError,
-    > for Rv64ZcbInstruction<Reg>
+    > for Rv32ZcbInstruction<Reg>
 where
-    Reg: Register<Type = u64>,
+    Reg: Register<Type = u32>,
     [(); Reg::N]:,
     Memory: VirtualMemory,
     PC: ProgramCounter<Reg::Type, Memory, CustomError>,
@@ -32,26 +34,26 @@ where
     ) -> Result<ControlFlow<()>, ExecutionError<Reg::Type, CustomError>> {
         match self {
             Self::CLbu { rd, rs1, uimm } => {
-                let addr = state.regs.read(rs1).wrapping_add(u64::from(uimm));
+                let addr = u64::from(state.regs.read(rs1).wrapping_add(u32::from(uimm)));
                 let value = state.memory.read::<u8>(addr)?;
-                state.regs.write(rd, u64::from(value));
+                state.regs.write(rd, u32::from(value));
             }
             Self::CLh { rd, rs1, uimm } => {
-                let addr = state.regs.read(rs1).wrapping_add(u64::from(uimm));
-                let value = i64::from(state.memory.read::<i16>(addr)?);
+                let addr = u64::from(state.regs.read(rs1).wrapping_add(u32::from(uimm)));
+                let value = i32::from(state.memory.read::<i16>(addr)?);
                 state.regs.write(rd, value.cast_unsigned());
             }
             Self::CLhu { rd, rs1, uimm } => {
-                let addr = state.regs.read(rs1).wrapping_add(u64::from(uimm));
+                let addr = u64::from(state.regs.read(rs1).wrapping_add(u32::from(uimm)));
                 let value = state.memory.read::<u16>(addr)?;
-                state.regs.write(rd, u64::from(value));
+                state.regs.write(rd, u32::from(value));
             }
             Self::CSb { rs1, rs2, uimm } => {
-                let addr = state.regs.read(rs1).wrapping_add(u64::from(uimm));
+                let addr = u64::from(state.regs.read(rs1).wrapping_add(u32::from(uimm)));
                 state.memory.write(addr, state.regs.read(rs2) as u8)?;
             }
             Self::CSh { rs1, rs2, uimm } => {
-                let addr = state.regs.read(rs1).wrapping_add(u64::from(uimm));
+                let addr = u64::from(state.regs.read(rs1).wrapping_add(u32::from(uimm)));
                 state.memory.write(addr, state.regs.read(rs2) as u16)?;
             }
             Self::CZextB { rd } => {
@@ -59,7 +61,7 @@ where
                 state.regs.write(rd, value);
             }
             Self::CSextB { rd } => {
-                let value = i64::from(state.regs.read(rd) as i8);
+                let value = i32::from(state.regs.read(rd) as i8);
                 state.regs.write(rd, value.cast_unsigned());
             }
             Self::CZextH { rd } => {
@@ -67,12 +69,8 @@ where
                 state.regs.write(rd, value);
             }
             Self::CSextH { rd } => {
-                let value = i64::from(state.regs.read(rd) as i16);
+                let value = i32::from(state.regs.read(rd) as i16);
                 state.regs.write(rd, value.cast_unsigned());
-            }
-            Self::CZextW { rd } => {
-                let value = state.regs.read(rd) & 0xffff_ffff;
-                state.regs.write(rd, value);
             }
             Self::CNot { rd } => {
                 let value = !state.regs.read(rd);
