+++ crates/execution/ab-riscv-interpreter/src/rv64/zce.rs
@@ -1,0 +1,3 @@
+//! RV64 Zce extension
+
+pub mod zcb;
