+++ crates/execution/ab-riscv-primitives/src/instructions/rv32.rs
@@ -5,6 +5,7 @@ pub mod c;
 pub mod m;
 #[cfg(test)]
 mod tests;
+pub mod zce;
 pub mod zk;
 
 use crate::instructions::Instruction;
