+++ crates/execution/ab-riscv-primitives/src/instructions/rv32/c.rs
@@ -1,4 +1,3 @@
 //! RV32 C extension
 
 pub mod zca;
-pub mod zcb;
