+++ crates/execution/ab-riscv-primitives/src/instructions/rv32/zce.rs
@@ -1,0 +1,3 @@
+//! RV32 Zce extension
+
+pub mod zcb;
