+++ crates/execution/ab-riscv-primitives/src/instructions/rv32/zce/zcb.rs
