+++ crates/execution/ab-riscv-primitives/src/instructions/rv64/c.rs
@@ -1,4 +1,3 @@
 //! RV64 C extension
 
 pub mod zca;
-pub mod zcb;
