+++ crates/execution/ab-riscv-primitives/src/instructions/rv64/zce/zcb.rs
