+++ crates/execution/ab-riscv-primitives/src/instructions/rv64/zce/zcb/tests.rs
@@ -1,7 +1,7 @@
 #![expect(clippy::identity_op, reason = "Test readability")]
 
 use crate::instructions::Instruction;
-use crate::instructions::rv64::c::zcb::Rv64ZcbInstruction;
+use crate::instructions::rv64::zce::zcb::Rv64ZcbInstruction;
 use crate::registers::general_purpose::Reg;
 
 /// Build a Q00 Zcb load/store: funct3=100, sub=bits\[12:10].
